% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\docType{package}
\name{CSSA-package}
\alias{CSSA-package}
\alias{CSSA}
\title{CSSA: Analysis and Simulation Tools for CRISPR-Cas9 Pooled Screens}
\description{
CSSA contains a number of functions to analyze and simulation pooled
CRISPR-Cas9-based screens. The simulators, represented by the CRISPRsim and
sortingsim functions, offer an easy but also highly customizable tool to
create data with guide- and gene-specific variables that can be randomly
assigned or specified by the user. The analysis tools currently contain
functions to calculate rate ratios, odds based on nonparametric guide
distribution, a function to combine genewise Z-values or odds, and a function
that calculates effect sizes of gene knockout and guide efficacy based on
rate ratios and duration of the experiment.
}
\section{CSSA functions}{

\describe{
\item{\code{\link{CRISPRsim}}}{Simulate a CRISPR-Cas9 pooled screen}
\item{\code{\link{radjust}}}{Calculate rate ratios restricted by confidence level}
\item{\code{\link{nestedradjust}}}{Calculate rate ratios of rate ratios restricted by confidence level}
\item{\code{\link{jar}}}{Rate ratios after adding an artificial number to all features}
\item{\code{\link{doublejar}}}{Rate ratios of rate ratios after adding an artificial number to all features}
\item{\code{\link{r2Z}}}{Convert rate ratios to Z-values}
\item{\code{\link{sumZ}}}{Calculate corrected summed Z-values per gene}
\item{\code{\link{getdeg}}}{Derive growth-modifying effect of gene knockout in pooled experiments}
\item{\code{\link{ess}}}{Get a list of essential genes or corresponding indices in the data set}
\item{\code{\link{noness}}}{Get a list of nonessential genes or corresponding indices in the data set}
}
}

\section{License}{

This package is licensed under GPL.
}

\author{
Jos B. Poell
}
