% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{doublejar}
\alias{doublejar}
\title{Rate ratios of rate ratios after adding an artificial number to all features}
\usage{
doublejar(mt1, wt1, mt0, wt0, n = 5, log = TRUE, normfun = "sum",
  normsubset)
}
\arguments{
\item{mt1}{Numeric vector. Feature data in test sample of line m}

\item{wt1}{Numeric vector. Feature data in test sample of line w}

\item{mt0}{Numeric vector. Feature data in control sample of line m}

\item{wt0}{Numeric vector. Feature data in control sample of line w}

\item{n}{Numeric. Specify how much is added to each value before calculating
rate ratios of rate ratios. Default = 5}

\item{log}{Logical or numeric. Specify whether to log-transform the rate
ratio, and with what base. If TRUE, uses log2. Default = TRUE}

\item{normfun}{Character string. Specify with which function to standardize
the data. Default = "sum"}

\item{normsubset}{Integer vector. Specify the indices of features that are to
be used in standardization}
}
\value{
Returns a numeric vector of the same length as the input vectors with
  the (log2-transformed) rate ratios of rate ratios after adding a specified
  number to all features.
}
\description{
doublejar adds a specified number to all features of all input vectors, after
which it calculates rate ratios of the rate ratios for all features. The
function was designed to prevent zeros in count data.
}
\seealso{
\code{\link{jar}}
}
\author{
Jos B. Poell
}
