% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{ess}
\alias{ess}
\title{Get a list of essential genes or corresponding indices in the data set}
\usage{
ess(guides, genes, trnc = "_.*")
}
\arguments{
\item{guides}{Character vector. List of guide names. Names of essential genes
are matched to the string before what is specified in the trnc argument.}

\item{genes}{Character vector. List of gene names. Names of essential genes 
are matched to the given gene names. Matches must be exact.}

\item{trnc}{Character. Regular expression pattern to isolate gene name from 
the guide name. Default = "_.*"}
}
\value{
Returns a character vector with names of essential genes, or a 
  numeric vector with indices representing essential genes in the given data 
  set.
}
\description{
ess couples reported essential genes to a data set. Without arguments, it 
returns the list of essential genes. Otherwise, it returns the indices 
corresponding to essential genes.
}
\note{
The list of essential genes was retrieved from
  http://hart-lab.org/downloads.
}
\seealso{
\code{\link{noness}}
}
\author{
Jos B. Poell
}
