% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{nestedradjust}
\alias{nestedradjust}
\title{Calculate rate ratios of rate ratios restricted by confidence level}
\usage{
nestedradjust(mtreads1, wtreads1, mtreads0, wtreads0, srmt1, srwt1, srmt0,
  srwt0, conf.level = 0.01, log = TRUE, belowxreads = FALSE)
}
\arguments{
\item{conf.level}{Numeric. Sets the confidence level of the rate ratio. If
FALSE, an unadjusted rate ratio will be returned. Default = 0.01}

\item{log}{Logical. Specify whether the rate ratio should be
log2-transformed. Default = TRUE}

\item{belowxreads}{Logical or numeric. Set threshold above which the
unadjusted rate ratio is returned. Default = FALSE}

\item{reads1}{Integer. Sequencing reads of feature in test sample}

\item{reads0}{Integer. Sequencing reads of feature in control sample}

\item{sumreads1}{Integer. Sum of all sequencing reads in test sample. Note
that this number is used for the "time base" in the poisson.test function
call. It can therefore also be the median reads, or a sum of a subset of
control genes.}

\item{sumreads2}{Integer. Sum of all sequencing reads in control sample. Note
that this number is used for the "time base" in the poisson.test function
call. It can therefore also be the median reads, or a sum of a subset of
control genes. Make sure to treat test and control equally.}
}
\value{
Returns the (log2-transformed) adjusted rate ratio.
}
\description{
nestedradjust is an extension of radjust. It is specifically applicable for
comparison of two samples with an independent time base.
}
\details{
The basic functionality is similar to \code{\link{radjust}}. The
  major difference lies in the time base factor. Because both samples have an
  independent origin, there is added uncertainty of this time base factor.
  This is where the nested aspect of this function comes in. Both the upper
  and lower confidence limit of the rate ratio at baseline are tested as time
  base. From the resulting two rate ratios, the smallest fold change ratio is
  returned.
}
\examples{
nestedradjust(30, 46, 100, 100, 21856736, 26730955, 20000000, 20000000)
nestedradjust(30, 46, 10, 10, 21856736, 26730955, 20000000, 20000000)

}
\seealso{
\code{\link{radjust}}
}
\author{
Jos B. Poell
}
