% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{noness}
\alias{noness}
\title{Get a list of nonessential genes or corresponding indices in the data set}
\usage{
noness(guides, genes, trnc = "_.*")
}
\arguments{
\item{guides}{Character vector. List of guide names. Names of nonessential genes
are matched to the string before what is specified in the trnc argument.}

\item{genes}{Character vector. List of gene names. Names of nonessential genes 
are matched to the given gene names. Matches must be exact.}

\item{trnc}{Character. Regular expression pattern to isolate gene name from 
the guide name. Default = "_.*"}
}
\value{
Returns a character vector with names of nonessential genes, or a 
  numeric vector with indices representing nonessential genes in the given data 
  set.
}
\description{
noness couples reported nonessential genes to a data set. Without arguments, it 
returns the list of nonessential genes. Otherwise, it returns the indices 
corresponding to nonessential genes.
}
\note{
The list of nonessential genes was retrieved from
  http://hart-lab.org/downloads.
}
\seealso{
\code{\link{ess}}
}
\author{
Jos B. Poell
}
