% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsscores.R
\name{odds2pq}
\alias{odds2pq}
\title{Calculate p-values and q-values for odds}
\usage{
odds2pq(odds, log = 10)
}
\arguments{
\item{odds}{Numeric vector. List of odds}

\item{log}{Integer or logical. Are the odds log-transformed or not? If not,
specify FALSE. If TRUE, the odds are presumed to be transformed by the
natural logarithm. Default = 10}
}
\value{
Returns a data frame with the odds, the p-values and the q-values
}
\description{
Returns a data frame with p-values and q-values for a vector of odds. When
using log odds, it keeps the p-values and q-values in the same log base.
}
\note{
The function basically calculates the probability of something being
  more extreme than the corresponding odds. Conversion uses the formula p =
  odds/(1+odds) or 1 - odds/(1+odds) if odds are higher than 1. The resulting
  value is multiplied by 2 to correct for 2-sided testing. For log odds, the
  absolute log value is subtracted from the log of 2 using the corresponding
  log base. Though not exact, this is a very close approximation with high or
  low odds, which is what this analysis is meant for.
}
\seealso{
\code{\link{oddscores}}, \code{\link{geneodds}}
}
\author{
Jos B. Poell
}
