% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{r2Z}
\alias{r2Z}
\title{Convert rate ratios to Z-values}
\usage{
r2Z(r, normsubset, method = "mean")
}
\arguments{
\item{r}{Character vector. List of rate ratios}

\item{normsubset}{Integer vector. Specify the indices of features that are to
be used in standardization}

\item{method}{Character. Specify what method to use for standardization.
Options are "median" and "mean". Default = "mean"}
}
\value{
Returns a numeric vector of the same length as the number of rate
  ratios
}
\description{
r2Z calculates the Z-values of rate ratios standardized to all rate ratios or
a specified normalization subset.
}
\details{
While not strictly required, it is recommended to input
  log-transformed rate ratios. Standardization is by default done using the
  mean and standard deviation: \code{(r-mean(r))/sd(r)}. By specifying method = "median", it does
  so using the median and median absolute deviations: \code{(r-median(r))/mad(r)}.
}
\note{
Functions in the CSSA package by default return log-transformed rate
  ratios. These can directly be used as input for this function.
}
\seealso{
\code{\link{sumZ}}
}
\author{
Jos B. Poell
}
