% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{radjust}
\alias{radjust}
\title{Calculate rate ratios restricted by confidence level}
\usage{
radjust(reads1, reads0, sumreads1, sumreads0, conf.level = 0.01, log = TRUE,
  belowxreads = FALSE)
}
\arguments{
\item{reads1}{Integer. Sequencing reads of feature in test sample}

\item{reads0}{Integer. Sequencing reads of feature in control sample}

\item{sumreads1}{Integer. Sum of all sequencing reads in test sample. Note
that this number is used for the "time base" in the poisson.test function
call. It can therefore also be the median reads, or a sum of a subset of
control genes.}

\item{conf.level}{Numeric. Sets the confidence level of the rate ratio. If
FALSE, an unadjusted rate ratio will be returned. Default = 0.01}

\item{log}{Logical. Specify whether the rate ratio should be
log2-transformed. Default = TRUE}

\item{belowxreads}{Logical or numeric. Set threshold above which the
unadjusted rate ratio is returned. Default = FALSE}

\item{sumreads2}{Integer. Sum of all sequencing reads in control sample. Note
that this number is used for the "time base" in the poisson.test function
call. It can therefore also be the median reads, or a sum of a subset of
control genes. Make sure to treat test and control equally.}
}
\value{
Returns the (log2-transformed) adjusted rate ratio.
}
\description{
radjust is designed to calculate rate ratios of sequencing results. It uses
the binomial distribution to calculate confidence intervals of rate ratios,
and returns the log2-transformed confidence limit that is closest to 0.
}
\details{
The core of this function utilizes the \code{\link{poisson.test}}
  function from the stats package, with \code{x = c(reads1, reads0)} and
  \code{T = c(sumreads1, sumreads0)}. If conf.level is not FALSE, the upper
  and lower confidence limits are log2-transformed and the value closest to 0
  is returned. If the log2-transformed upper and lower limit have opposite
  signs, 0 is returned.
}
\note{
If conf.level is set to 0, it will be interpreted as FALSE. This is
  intended, as radjust will then return an unadjusted rate ratio estimate,
  instead of giving an error.
}
\examples{
radjust(30, 46, 21856736, 26730955)
radjust(30, 46, 21856736, 26730955, conf.level = 0.8)
radjust(30, 46, 21856736, 26730955, belowxreads = 30)
radjust(30, 46, 21856736, 26730955, log = FALSE)
t1 <- c(658,2464,45,355677,568)
t0 <- c(234,6334,74,343346,533)
mapply(radjust, t1, t0, sum(t1), sum(t0))

}
\seealso{
\code{\link{nestedradjust}}
}
\author{
Jos B. Poell
}
