% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrep_degrep.R
\name{rrep}
\alias{rrep}
\title{Calculate rate ratios with standard error of count data based on replicates}
\usage{
rrep(t1, t0, paired = TRUE, normfun = "sum", normsubset,
  rstat = "summed")
}
\arguments{
\item{t1}{Matrix or data frame, with rows representing features and columns
representing replicates of measurements at t1 (or treated).}

\item{t0}{Matrix or data frame, with rows representing features and columns
representing replicates of measurements at t0 (or untreated).}

\item{paired}{Logical. Are measurements paired? Default = TRUE}

\item{normfun}{Character string. Specify with which function to standardize
the data. Default = "sum"}

\item{normsubset}{Integer vector. Specify the indices of features that are to
be used in standardization}

\item{rstat}{Character string. Specify whether rate ratios are calculated
over the sum of the counts in replicates or as the mean of the log2 rate
ratios. Default = "summed"}
}
\value{
Returns a data frame with the log2-transformed rate ratio and
  corresponding standard error of each feature.
}
\description{
rrep calculates rate ratios and corresponding standard errors of features
based on the presence of replicates. The function uses the compound variance
of the rate ratio of replicates as well as the variance associated with read
depth (i.e. counts). Rate ratio and standard error are expressed as
log2-values.
}
\details{
This function combines the confidence based on replicate
  measurements with the confidence based on counts (e.g. read depth).
  Utilizing replicates to assess confidence in point estimates of individual
  features is commonplace in many analyses. However, in data sets with many
  features just by chance some features will have measurements that lie very
  close together. By adding the variance based on the count data, spurious
  findings are greatly reduced, especially when counts are low. Variance of
  count data on a log2-transformed scale is approximated with the formula
  \code{1/(log(2)^2*count)}
}
\note{
If the data contain zeros, a pseudocount of 1/replicates is added to
  all counts.
}
\examples{
  set.seed(1000)
  c0 <- rbind(rpois(3, 60), rpois(3, 5), rpois(3, 10000000))
  c1 <- rbind(rpois(3, 20), rpois(3, 20), rpois(3, 10000000))
  rrep(c1, c0, paired = FALSE)
  c01 <- rbind(rep(60, 3), rep(5, 3), rep(10000000, 3))
  c11 <- rbind(rep(20, 3), rep(20, 3), rep(10000000,3))
  rrep(c11, c01)
  c04 <- 4*c01
  c14 <- 4*c11
  rrep(c14,c04)

}
\seealso{
\code{\link{degrep}}, \code{\link{CRISPRsim}}, \code{\link{ess}},
  \code{\link{noness}}
}
\author{
Jos B. Poell
}
