% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{sumZ}
\alias{sumZ}
\title{Calculate corrected summed Z-values per gene}
\usage{
sumZ(guides, Z)
}
\arguments{
\item{guides}{Character vector. List of guides}

\item{Z}{Numeric vector. Z-values of guides}
}
\value{
Returns a numeric vector of the same length as the number of genes
  with the corrected summed Z-value.
}
\description{
sumZ adds Z-values of all guides per gene, then corrects each value by
dividing by the square root of the number of guides representing the gene
}
\note{
Z-values can be based on the entire population, but also on a
  population of control genes. See examples below
}
\examples{
ut <- CRISPRsim(5000, 4, a = c(3,3), allseed = 1)
tr <- CRISPRsim(5000, 4, a = c(3,3), e = TRUE, allseed = 1)
cgi <- tr$d > -0.05 & tr$d < 0.05 & tr$e > -0.05 & tr$e < 0.05
r0 <- jar(ut$t6, ut$t0)
r1 <- jar(tr$t6, ut$t6)
Z0 <- (r0-mean(r0[cgi]))/sd(r0[cgi])
Z1 <- (r1-mean(r1[cgi]))/sd(r1[cgi])
sumZ0 <- sumZ(tr$guides, Z0)
sumZ1 <- sumZ(tr$guides, Z1)
reald <- rle(tr$d)$values
reale <- rle(tr$e)$values
plot(reald, sumZ0)
plot(reale, sumZ1)

}
\seealso{
\code{\link{CRISPRsim}}, \code{\link{jar}}, \code{\link{radjust}}
}
\author{
Jos B. Poell
}
