% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{nestedradjust}
\alias{nestedradjust}
\title{Calculate rate ratios of rate ratios restricted by confidence level}
\usage{
nestedradjust(mt1, wt1, mt0, wt0, conf.level = 0.01, normfun = "sum",
  normsubset, log = TRUE, belowxreads = FALSE)
}
\arguments{
\item{mt1}{Integer vector. Raw sequencing reads in test sample, t1}

\item{wt1}{Integer vector. Raw sequencing reads in control sample, t1}

\item{mt0}{Integer vector. Raw sequencing reads in test sample, t0}

\item{wt0}{Integer vector. Raw sequencing reads in control sample, t0}

\item{conf.level}{Numeric. Sets the confidence level of the rate ratio. If
FALSE, an unadjusted rate ratio will be returned. Default = 0.01}

\item{normfun}{Character string. Specify with which function to standardize
the data. Default = "sum"}

\item{normsubset}{Integer vector. Specify the indices of features that are to
be used in standardization}

\item{log}{Logical. Specify whether the rate ratio should be
log2-transformed. If TRUE, uses log2. Default = TRUE}

\item{belowxreads}{Logical or numeric. Set threshold above which the
unadjusted rate ratio is returned. Default = FALSE}
}
\value{
Returns the (log2-transformed) adjusted rate ratio.
}
\description{
nestedradjust is an extension of radjust. It is specifically applicable for
comparison of two samples with an independent time base.
}
\details{
The basic functionality is similar to \code{\link{radjust}}. The
  major difference lies in the time base factor. Because both samples have an
  independent origin, there is added uncertainty of this time base factor.
  This is where the nested aspect of this function comes in. Both the upper
  and lower confidence limit of the rate ratio at baseline are tested as time
  base. From the resulting two rate ratios, the smallest fold change ratio is
  returned. It therefore estimates a conservative rate ratio.
}
\examples{
wt <- CRISPRsim(genes = 10, guides = 4, a = 3, allseed = 1, t0seed = 2)
mt <- CRISPRsim(genes = 10, guides = 4, a = 3, e = TRUE, f = jitter(wt$f),
                allseed = 1, repseed = 2)
nestedradjust(mt$t3, wt$t3, mt$t0, wt$t0)

}
\seealso{
\code{\link{radjust}}, \code{\link{CRISPRsim}}
}
\author{
Jos B. Poell
}
