% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSSA.R
\name{radjust}
\alias{radjust}
\title{Calculate rate ratios restricted by confidence level}
\usage{
radjust(t1, t0, conf.level = 0.01, normfun = "sum", normsubset,
  log = TRUE, belowxreads = FALSE)
}
\arguments{
\item{t1}{Integer vector. Raw sequencing reads in test sample}

\item{t0}{Integer vector. Raw sequencing reads in control sample}

\item{conf.level}{Numeric. Sets the confidence level of the rate ratio. If
FALSE, an unadjusted rate ratio will be returned. Default = 0.01}

\item{normfun}{Character string. Specify with which function to standardize
the data. Default = "sum"}

\item{normsubset}{Integer vector. Specify the indices of features that are to
be used in standardization}

\item{log}{Logical or numeric. Specify whether to log-transform the rate
ratio, and with what base. If TRUE, uses log2. Default = TRUE}

\item{belowxreads}{Logical or numeric. Set threshold above which the
unadjusted rate ratio is returned. Default = FALSE}
}
\value{
Returns the (log2-transformed) adjusted rate ratio.
}
\description{
radjust is designed to calculate rate ratios of sequencing results. It uses
the binomial distribution to calculate confidence intervals of rate ratios,
and returns the log2-transformed confidence limit that is closest to 0.
}
\details{
The core of this function utilizes the \code{\link{poisson.test}}
  function from the stats package, with \code{x = c(t1, t0)} and
  \code{T = c(sumreads1, sumreads0)}. If conf.level is not FALSE, the upper
  and lower confidence limits are log2-transformed and the value closest to 0
  is returned. If the log2-transformed upper and lower limit have opposite
  signs, 0 is returned.
}
\note{
If conf.level is set to 0, it will be interpreted as FALSE. This is
  intended, as radjust will then return an unadjusted rate ratio estimate,
  instead of giving an error.
}
\examples{
ut <- CRISPRsim(200, 4, a = c(3,3), allseed = 1)
tr <- CRISPRsim(200, 4, a = c(3,3), e = TRUE, allseed = 1)
cgi <- tr$d > -0.05 & tr$d < 0.05 & tr$e > -0.05 & tr$e < 0.05
r0 <- radjust(ut$t3, ut$t0, belowxreads = 300, normsubset = cgi)
r1 <- radjust(tr$t3, ut$t3, belowxreads = 300, normsubset = cgi)
hcr0 <- radjust(ut$t3, ut$t0, conf.level = 0.8, normsubset = cgi)
hcr1 <- radjust(tr$t3, ut$t3, conf.level = 0.8, normsubset = cgi)
plot(r0, hcr0)
abline(0, 1)
plot(r1, hcr1)
abline(0, 1)

}
\seealso{
\code{\link{nestedradjust}}
}
\author{
Jos B. Poell
}
